pyproject.toml
[project]
name = "startup-validator-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.21",
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.45",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
]
